Added the time class + heaps.
I intend to standardise time and rules to be objects, rather than my
previous ideas of interpretations of strings. I also standardised
constants and functions that were needed in multiple parts of the
project. I think currently some of the stuff in time might not run, but,
it is all going to be largely overhauled anyway, so, it shoudl be fine.