Did many a thing, including make_week
Removed the gap file/class, which had already been replaced with
TimeInterval.

Filled in make_week, the incredibly important function which generates a
filled week object from a list of rules passed to it.

Did some other stuff.

Added an is_zero function to TimeInterval, so you can use intervals even
as no interval.

Removed the stuff which didn't need to be in the readme.

Added a to_do list, so I don't forget stuff.