Display restructure + removed event class
Display restructure to allow for text display initially.

Event was removed as it would have made it over complicated to read from
files. I have simply now made it a string, that should be all that is
needed initially.

Perhaps even, there will simply be a symbol in the string which
separates the name from the description.